requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
argon2-cffi==23.1.0
    # via minio
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
certifi==2024.8.30
    # via
    #   minio
    #   requests
cffi==1.17.1
    # via argon2-cffi-bindings
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.4.0
    # via requests
cupy-cuda12x==13.3.0
    # via -r requirements.in
distlib==0.3.9
    # via virtualenv
fastrlock==0.8.3
    # via cupy-cuda12x
filelock==3.16.1
    # via
    #   torch
    #   virtualenv
fsspec==2024.6.1
    # via torch
identify==2.6.1
    # via pre-commit
idna==3.10
    # via requests
iniconfig==2.0.0
    # via pytest
jinja2==3.1.4
    # via
    #   memray
    #   torch
joblib==1.4.2
    # via scikit-learn
liac-arff==2.5.0
    # via openml
linkify-it-py==2.0.3
    # via markdown-it-py
markdown-it-py==3.0.0
    # via
    #   mdit-py-plugins
    #   rich
    #   textual
markupsafe==3.0.2
    # via jinja2
mdit-py-plugins==0.4.2
    # via markdown-it-py
mdurl==0.1.2
    # via markdown-it-py
memray==1.15.0
    # via -r requirements.in
minio==7.2.10
    # via openml
mpmath==1.3.0
    # via sympy
networkx==3.4.2
    # via torch
nodeenv==1.9.1
    # via
    #   pre-commit
    #   pyright
numpy==2.1.2
    # via
    #   -r requirements.in
    #   cupy-cuda12x
    #   openml
    #   pandas
    #   scikit-learn
    #   scipy
openml==0.15.0
    # via -r requirements.in
packaging==24.1
    # via
    #   openml
    #   pytest
pandas==2.2.3
    # via openml
platformdirs==4.3.6
    # via
    #   textual
    #   virtualenv
pluggy==1.5.0
    # via pytest
pre-commit==4.0.1
    # via -r requirements.in
psutil==6.1.0
    # via -r requirements.in
pyarrow==18.0.0
    # via openml
pycparser==2.22
    # via cffi
pycryptodome==3.21.0
    # via minio
pygments==2.18.0
    # via rich
pyright==1.1.384
    # via -r requirements.in
pytest==8.3.3
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via
    #   openml
    #   pandas
pytz==2024.2
    # via pandas
pyyaml==6.0.2
    # via pre-commit
regex==2024.11.6
    # via
    #   -r requirements.in
    #   tiktoken
requests==2.32.3
    # via
    #   openml
    #   tiktoken
rich==13.9.4
    # via
    #   memray
    #   textual
ruff==0.7.0
    # via -r requirements.in
scikit-learn==1.5.2
    # via
    #   -r requirements.in
    #   openml
scipy==1.14.1
    # via
    #   openml
    #   scikit-learn
setuptools==75.8.0
    # via torch
six==1.16.0
    # via python-dateutil
sympy==1.13.1
    # via torch
textual==1.0.0
    # via memray
threadpoolctl==3.5.0
    # via scikit-learn
tiktoken==0.8.0
    # via -r requirements.in
torch==2.5.0
    # via -r requirements.in
tqdm==4.66.5
    # via
    #   -r requirements.in
    #   openml
typing-extensions==4.12.2
    # via
    #   minio
    #   pyright
    #   textual
    #   torch
tzdata==2024.2
    # via pandas
uc-micro-py==1.0.3
    # via linkify-it-py
urllib3==2.2.3
    # via
    #   minio
    #   requests
virtualenv==20.26.6
    # via pre-commit
xmltodict==0.14.2
    # via openml
